requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
captcha==0.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.19
django-cors-headers==4.0.0
django-filter==23.2
django-guardian==2.4.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.6.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.3
Pillow==9.5.0
pip==23.1.2
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
setuptools==65.5.0
snowflake-id==0.0.2
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0